Fix compilation with -Wsign-conversion (#1367)
* Fix compilation with -Wsign-conversion

For tests with GCC and Clang

Fixes https://github.com/ETLCPP/etl/issues/632

* Fix support of negative Id in type_lookup.h: type_from_id

* Fix element access arithmetic in atomic_gcc_sync.h fetch_add/fetch_sub

* Fix rounded_integral_division.h: divide_round_half_odd(): direction is always 1

+++ include/etl/array.h
@@ -496,13 +496,13 @@ namespace etl
       iterator p = to_iterator(position);
       iterator result(p);
 
-      size_t source_size       = etl::distance(first, last);
-      size_t destination_space = etl::distance(position, cend());
+      size_t source_size       = static_cast<size_t>(etl::distance(first, last));
+      size_t destination_space = static_cast<size_t>(etl::distance(position, cend()));
 
       // Do we need to move anything?
       if (source_size < destination_space)
       {
-        size_t length = SIZE - (etl::distance(begin(), p) + source_size);
+        size_t length = SIZE - (static_cast<size_t>(etl::distance(begin(), p)) + source_size);
         etl::move_backward(p, p + length, end());
       }
