+++ test/test_crc8_nrsc5.cpp
@@ -77,7 +77,7 @@ namespace
 
       for (size_t i = 0UL; i < data.size(); ++i)
       {
-        crc_calculator.add(data[i]);
+        crc_calculator.add(static_cast<uint8_t>(data[i]));
       }
 
       uint8_t crc = crc_calculator;
@@ -168,7 +168,7 @@ namespace
 
       for (size_t i = 0UL; i < data.size(); ++i)
       {
-        crc_calculator.add(data[i]);
+        crc_calculator.add(static_cast<uint8_t>(data[i]));
       }
 
       uint8_t crc = crc_calculator;
@@ -259,7 +259,7 @@ namespace
 
       for (size_t i = 0UL; i < data.size(); ++i)
       {
-        crc_calculator.add(data[i]);
+        crc_calculator.add(static_cast<uint8_t>(data[i]));
       }
 
       uint8_t crc = crc_calculator;
