+++ test/test_span_dynamic_extent.cpp
@@ -966,7 +966,7 @@ namespace
 
       View::circular_iterator sci = view.begin_circular();
 
-      for (int i = 0; i < 20; ++i)
+      for (size_t i = 0U; i < 20U; ++i)
       {
         CHECK_EQUAL(expected[i], *++sci);
       }
@@ -983,7 +983,7 @@ namespace
 
       View::circular_iterator sci = subspan.begin_circular();
 
-      for (int i = 0; i < 20; ++i)
+      for (size_t i = 0U; i < 20U; ++i)
       {
         CHECK_EQUAL(expected[i], *++sci);
       }
@@ -999,7 +999,7 @@ namespace
 
       View::circular_iterator sci = view.begin_circular();
 
-      for (int i = 0; i < 20; ++i)
+      for (size_t i = 0U; i < 20U; ++i)
       {
         CHECK_EQUAL(expected[i % ETL_OR_STD17::size(expected)], *sci++);
       }
@@ -1016,7 +1016,7 @@ namespace
 
       View::circular_iterator sci = subspan.begin_circular();
 
-      for (int i = 0; i < 20; ++i)
+      for (size_t i = 0U; i < 20U; ++i)
       {
         CHECK_EQUAL(expected[i], *sci++);
       }
@@ -1032,7 +1032,7 @@ namespace
 
       View::reverse_circular_iterator sci = view.rbegin_circular();
 
-      for (int i = 0; i < 20; ++i)
+      for (size_t i = 0U; i < 20U; ++i)
       {
         CHECK_EQUAL(expected[i % ETL_OR_STD17::size(expected)], *++sci);
       }
@@ -1049,7 +1049,7 @@ namespace
 
       etl::span<int, 10U>::reverse_circular_iterator sci = subspan.rbegin_circular();
 
-      for (int i = 0; i < 20; ++i)
+      for (size_t i = 0U; i < 20U; ++i)
       {
         CHECK_EQUAL(expected[i % ETL_OR_STD17::size(expected)], *++sci);
       }
@@ -1065,7 +1065,7 @@ namespace
 
       View::reverse_circular_iterator sci = view.rbegin_circular();
 
-      for (int i = 0; i < 20; ++i)
+      for (size_t i = 0U; i < 20U; ++i)
       {
         CHECK_EQUAL(expected[i % ETL_OR_STD17::size(expected)], *sci++);
       }
@@ -1082,7 +1082,7 @@ namespace
 
       View::reverse_circular_iterator sci = subspan.rbegin_circular();
 
-      for (int i = 0; i < 20; ++i)
+      for (size_t i = 0U; i < 20U; ++i)
       {
         CHECK_EQUAL(expected[i % ETL_OR_STD17::size(expected)], *sci++);
       }
@@ -1100,7 +1100,7 @@ namespace
       {
         View::circular_iterator sci = view.begin_circular();
 
-        for (int i = 0; i < 20; i += step)
+        for (size_t i = 0U; i < 20U; i += static_cast<size_t>(step))
         {
           CHECK_EQUAL(expected[i % 10], *sci);
           sci += step;
@@ -1120,7 +1120,7 @@ namespace
       {
         View::circular_iterator sci = view.begin_circular();
 
-        for (int i = 0; i < 20; i += step)
+        for (size_t i = 0U; i < 20U; i += static_cast<size_t>(step))
         {
           CHECK_EQUAL(expected[i % 10], *sci);
           sci = sci + step;
@@ -1140,7 +1140,7 @@ namespace
       {
         View::circular_iterator sci = view.begin_circular();
 
-        for (int i = 0; i < 20; i += step)
+        for (size_t i = 0U; i < 20U; i += static_cast<size_t>(step))
         {
           CHECK_EQUAL(expected[i % 10], *sci);
           sci -= step;
@@ -1160,7 +1160,7 @@ namespace
       {
         View::circular_iterator sci = view.begin_circular();
 
-        for (int i = 0; i < 20; i += step)
+        for (size_t i = 0U; i < 20U; i += static_cast<size_t>(step))
         {
           CHECK_EQUAL(expected[i % 10], *sci);
           sci = sci - step;
