+++ test/test_unordered_set.cpp
@@ -142,12 +142,12 @@ namespace
 
     size_t operator ()(const char* s) const
     {
-      return std::accumulate(s, s + etl::strlen(s), 0);
+      return std::accumulate(s, s + etl::strlen(s), size_t(0));
     }
 
     size_t operator ()(const std::string& s) const
     {
-      return std::accumulate(s.begin(), s.end(), 0);
+      return std::accumulate(s.begin(), s.end(), size_t(0));
     }
   };
 
@@ -853,7 +853,7 @@ namespace
       CHECK_CLOSE(0.0, data.load_factor(), 0.01);
 
       // Half the buckets used.
-      data.assign(initial_data.begin(), initial_data.begin() + (initial_data.size() / 4));
+      data.assign(initial_data.begin(), initial_data.begin() + static_cast<ptrdiff_t>(initial_data.size() / 4));
       CHECK_CLOSE(0.4, data.load_factor(), 0.01);
 
       // All of the buckets used.
@@ -922,7 +922,7 @@ namespace
         // Force hash collisions
         size_t operator()(int key) const
         {
-          return key % 4;
+          return static_cast<size_t>(key % 4);
         }
       };
